[Refactor] Move adstack-overflow registry + diagnose classifier + launch snapshot from Program into AdStackCache
