+++ quadrants/program/program.cpp
@@ -43,7 +43,8 @@
 namespace quadrants::lang {
 std::atomic<int> Program::num_instances_;
 
-Program::Program(Arch desired_arch) : snode_rw_accessors_bank_(this), adstack_cache_(std::make_unique<AdStackCache>()) {
+Program::Program(Arch desired_arch)
+    : snode_rw_accessors_bank_(this), adstack_cache_(std::make_unique<AdStackCache>(this)) {
   QD_TRACE("Program initializing...");
 
   // For performance considerations and correctness of QuantFloatType operations, we force floating-point operations to
@@ -168,7 +169,7 @@ void Program::launch_kernel(const CompiledKernelData &compiled_kernel_data, Laun
   // `set_host_accessible_ndarray_ptrs` overwrites `array_ptrs[(arg_id, DATA_PTR_POS)]` from the original
   // `DeviceAllocation *` handle to a raw host pointer; capturing earlier preserves the handle so the diagnose
   // path's `Device::map(*alloc, ...)` lookup is uniform across CPU / CUDA / AMDGPU / Vulkan / Metal.
-  capture_diagnose_snapshot(ctx);
+  adstack_cache_->capture_diagnose_snapshot(ctx);
   num_offloaded_tasks_on_last_call_ = compiled_kernel_data.num_tasks();
   program_impl_->get_kernel_launcher().launch_kernel(compiled_kernel_data, ctx);
   if (compile_config().debug && arch_uses_llvm(compiled_kernel_data.arch())) {
@@ -249,259 +250,6 @@ void Program::synchronize_and_assert() {
 
 void Program::check_adstack_overflow_and_assert() {
   program_impl_->check_adstack_overflow_and_assert();
-}
-
-uint32_t Program::register_adstack_sizing_info(const void *identity_key,
-                                               const std::string &kernel_name,
-                                               int task_id_in_kernel,
-                                               std::vector<int> allocated_max_sizes,
-                                               std::vector<SerializedSizeExpr> size_exprs) {
-  std::lock_guard<std::mutex> lk(adstack_sizing_info_registry_mutex_);
-  // Idempotent re-registration: same `identity_key` yields the same id across re-compiles and updates
-  // the entry's metadata + size_exprs in place. The key is just an opaque dedup token - the registry
-  // never dereferences it; all data needed by the diagnose path is copied into the entry below. This
-  // sidesteps the lifetime issue we hit with deref-based registries (the codegen-time
-  // `&current_task->ad_stack` is freed by the launcher's `current_task = nullptr` after a by-value
-  // `offloaded_tasks.push_back(*current_task)`, leaving the registered pointer staring at poisoned
-  // unique_ptr heap; verified via instrumentation: `0xAAAAAAAAAAAAAAAD` marker on the next deref).
-  auto it = adstack_sizing_info_id_by_ptr_.find(identity_key);
-  if (it != adstack_sizing_info_id_by_ptr_.end()) {
-    auto &entry = adstack_sizing_info_registry_[it->second];
-    entry.kernel_name = kernel_name;
-    entry.task_id_in_kernel = task_id_in_kernel;
-    entry.allocated_max_sizes = std::move(allocated_max_sizes);
-    entry.size_exprs = std::move(size_exprs);
-    return it->second;
-  }
-  uint32_t id = static_cast<uint32_t>(adstack_sizing_info_registry_.size());
-  AdStackSizingInfoEntry entry;
-  entry.identity_key = identity_key;
-  entry.kernel_name = kernel_name;
-  entry.task_id_in_kernel = task_id_in_kernel;
-  entry.allocated_max_sizes = std::move(allocated_max_sizes);
-  entry.size_exprs = std::move(size_exprs);
-  adstack_sizing_info_registry_.push_back(std::move(entry));
-  adstack_sizing_info_id_by_ptr_.emplace(identity_key, id);
-  return id;
-}
-
-void Program::update_adstack_sizing_info_size_exprs(uint32_t id, std::vector<SerializedSizeExpr> size_exprs) {
-  std::lock_guard<std::mutex> lk(adstack_sizing_info_registry_mutex_);
-  if (id == 0 || id >= adstack_sizing_info_registry_.size()) {
-    return;
-  }
-  adstack_sizing_info_registry_[id].size_exprs = std::move(size_exprs);
-}
-
-std::optional<Program::AdStackSizingInfoEntry> Program::lookup_adstack_sizing_info(uint32_t id) const {
-  std::lock_guard<std::mutex> lk(adstack_sizing_info_registry_mutex_);
-  if (id == 0 || id >= adstack_sizing_info_registry_.size()) {
-    return std::nullopt;
-  }
-  return adstack_sizing_info_registry_[id];
-}
-
-std::string Program::diagnose_adstack_overflow_message(uint32_t task_id) const {
-  return diagnose_adstack_overflow(task_id).message;
-}
-
-Program::AdStackOverflowDiagnosis Program::diagnose_adstack_overflow(uint32_t task_id) const {
-  std::string identity_block;
-  std::string disambiguation_block;
-  // Cause classifier: when the synchronous re-run produces required > allocated for ANY stack, the
-  // most likely cause is an untracked tensor mutation (DLPack-bypass etc.). When all required <=
-  // allocated, the pre-pass undersized the bound (Quadrants bug). When we cannot re-evaluate
-  // (ndarray-bound size_expr; no launch ctx in hand) we fall through to the static dual-cause body.
-  enum class Cause { Unknown, DLPackBypass, QuadrantsBug };
-  Cause cause = Cause::Unknown;
-
-  if (task_id != 0) {
-    auto entry_opt = lookup_adstack_sizing_info(task_id);
-    if (entry_opt.has_value()) {
-      const auto &entry = *entry_opt;
-      identity_block = "  Offending task: kernel `" + entry.kernel_name + "` offload task #" +
-                       std::to_string(entry.task_id_in_kernel) + "; per-stack allocated max_size = [";
-      for (size_t i = 0; i < entry.allocated_max_sizes.size(); ++i) {
-        if (i != 0) {
-          identity_block += ", ";
-        }
-        identity_block += std::to_string(entry.allocated_max_sizes[i]);
-      }
-      identity_block += "].\n";
-
-      // Synchronous sizer rerun: walk each stack's `SerializedSizeExpr` and evaluate against the live
-      // host / SNode state. Skip stacks whose tree contains an `ExternalTensorShape` or
-      // `ExternalTensorRead` leaf - those need the per-launch `LaunchContextBuilder` which is no longer
-      // available at this point. The disambiguation is best-effort: if every stack's tree is host-
-      // resolvable we get a precise classification; otherwise we report what we have and the caller
-      // falls back to the static dual-cause hint. The eval path itself goes through
-      // `SNodeRwAccessorsBank` for `FieldLoad` leaves, which dispatches its own reader kernel; the
-      // diagnose path only fires on the error branch so the extra dispatches are not on any hot path.
-      //
-      // Lifetime: the registry's `ad_stack_ptr` is rebound by each backend launcher via
-      // `Program::set_adstack_sizing_info_pointer` to the post-`push_back` address in
-      // `offloaded_tasks` (LLVM) or `task_attribs.ad_stack` (SPIR-V), which is stable for the cached
-      // kernel's lifetime. The codegen-time pointer is rebound on the FIRST launch; before that the
-      // entry's `ad_stack_ptr` still points at freed unique_ptr heap, so the deref below is gated on a
-      // sentinel-tag check the launcher writes when it rebinds (see `set_adstack_sizing_info_pointer`).
-      // The registry entry's `size_exprs` are heap-owned copies of the per-alloca size-expression
-      // trees, copied at `register_adstack_sizing_info` time. Walking them is safe regardless of the
-      // backend's storage layout (LLVM `AdStackSizingInfo::size_exprs` parallel-to-allocas vs SPIR-V
-      // `AdStackSizingAttribs::allocas[i].size_expr` inline) and immune to launcher-side struct
-      // moves that would invalidate raw pointers into the original kernel data.
-      if (!entry.size_exprs.empty()) {
-        std::vector<int64_t> required_sizes;
-        std::vector<bool> required_known;
-        size_t any_grew = 0;
-        size_t any_unknown = 0;
-        size_t total = std::min(entry.size_exprs.size(), entry.allocated_max_sizes.size());
-        for (size_t i = 0; i < total; ++i) {
-          const auto &expr = entry.size_exprs[i];
-          bool host_resolvable = true;
-          for (const auto &node : expr.nodes) {
-            auto k = static_cast<SizeExpr::Kind>(node.kind);
-            if (k == SizeExpr::Kind::ExternalTensorShape || k == SizeExpr::Kind::ExternalTensorRead) {
-              host_resolvable = false;
-              break;
-            }
-          }
-          int64_t v = -1;
-          if (host_resolvable && !expr.nodes.empty()) {
-            // `evaluate_adstack_size_expr` reads the live SNode state via `SNodeRwAccessorsBank` and const-folds the
-            // rest in plain C++. `ctx == nullptr` is safe because every leaf we kept is host-resolvable; ETS / ETR
-            // are the only kinds that touch ctx and we filtered them out.
-            SizeExprLaunchScope scope;
-            v = evaluate_adstack_size_expr(expr, const_cast<Program *>(this), nullptr);
-          } else if (!expr.nodes.empty()) {
-            // Tree contains `ExternalTensorRead` / `ExternalTensorShape` leaves. Try the diagnose-time evaluator
-            // which resolves them against the captured launch snapshot (`Device::map`-based ndarray reads, see
-            // `Program::DiagnoseLaunchSnapshot`'s comment for the design rationale). On any leaf-resolution
-            // failure (no snapshot yet, allocation cannot be mapped, unsupported dtype) the helper returns -1
-            // and we fall through to the `?` placeholder.
-            int64_t diag = evaluate_adstack_size_expr_for_diagnose(expr, const_cast<Program *>(this));
-            if (diag >= 0) {
-              v = diag;
-            }
-          }
-          required_sizes.push_back(v);
-          required_known.push_back(!expr.nodes.empty() && v >= 0);
-          if (required_known.back() && static_cast<size_t>(v) > entry.allocated_max_sizes[i]) {
-            ++any_grew;
-          }
-          if (!required_known.back()) {
-            ++any_unknown;
-          }
-        }
-        if (any_grew > 0) {
-          cause = Cause::DLPackBypass;
-        } else if (any_unknown == 0 && total > 0) {
-          cause = Cause::QuadrantsBug;
-        }
-        // Only print the rerun line when at least one stack's bound resolves to a real value. With every
-        // leaf unresolved the line would be `required = [?, ?, ...]` which adds zero signal beyond the
-        // dual-cause body that follows; the omission keeps the message focused on actionable content.
-        if (any_unknown < total) {
-          disambiguation_block = "  Synchronous sizer rerun: required max_size = [";
-          for (size_t i = 0; i < required_sizes.size(); ++i) {
-            if (i != 0) {
-              disambiguation_block += ", ";
-            }
-            if (required_known[i]) {
-              disambiguation_block += std::to_string(required_sizes[i]);
-            } else {
-              disambiguation_block += "?";
-            }
-          }
-          disambiguation_block += "].";
-          if (any_unknown > 0) {
-            // Only annotate the placeholder if at least one stack actually fell back to it.
-            disambiguation_block +=
-                " (`?` = sizer rerun could not resolve this stack's bound against the captured "
-                "launch state).";
-          }
-          disambiguation_block += "\n";
-        }
-      }
-    }
-  }
-
-  std::string body;
-  if (cause == Cause::DLPackBypass) {
-    body =
-        "Cause (sync sizer rerun): a tensor backing a data-dependent loop bound was mutated outside "
-        "Quadrants's tracking - typically a DLPack zero-copy mutation through a torch tensor sharing "
-        "storage with a Quadrants ndarray, or a raw pointer write through a non-torch DLPack consumer. "
-        "The cached adstack capacity was sized against the value before the mutation. Recovery: route "
-        "the mutation through Quadrants APIs (`Ndarray.write` / `fill` / kernel writes) so the cache "
-        "invalidates correctly, OR set a generous initial cap if a workload-change milestone genuinely "
-        "grew capacity. Restart the iteration / training loop from a clean state.\n";
-  } else if (cause == Cause::QuadrantsBug) {
-    body =
-        "Cause (sync sizer rerun): the freshly-computed required size does not exceed the allocated "
-        "size for any stack - this is a Quadrants bug. The pre-pass resolved the alloca to a bound "
-        "tighter than the actual runtime push count: either the enclosing loop shape is outside the "
-        "current `SizeExpr` grammar, or the Bellman-Ford analyzer undercounted the forward-pass "
-        "accumulation. Please file with the kernel IR (`QD_DUMP_IR=1`).\n";
-  } else {
-    body =
-        "Two possible causes (synchronous sizer rerun was not conclusive - some `SizeExpr` trees "
-        "depend on ndarray contents that are not host-resolvable without a per-launch context, or the "
-        "task-id slot was empty so the registry pointer could not be confirmed live):\n"
-        "  1. A tensor backing a data-dependent loop bound was mutated outside Quadrants's tracking "
-        "(typically a DLPack zero-copy mutation through a torch tensor sharing storage with a "
-        "Quadrants ndarray, or a raw pointer write through a non-torch DLPack consumer). The cached "
-        "adstack capacity was sized against the value before the mutation. Recovery: route the "
-        "mutation through Quadrants APIs (`Ndarray.write` / `fill` / kernel writes) so the cache "
-        "invalidates correctly, OR set a generous initial cap if a workload-change milestone "
-        "genuinely grew capacity. Restart the iteration / training loop from a clean state.\n"
-        "  2. (Quadrants bug) the pre-pass resolved the alloca to a bound tighter than the actual "
-        "runtime push count - the enclosing loop shape is outside the current `SizeExpr` grammar, or "
-        "the Bellman-Ford analyzer undercounted the forward-pass accumulation. Please file with the "
-        "kernel IR (`QD_DUMP_IR=1`).\n";
-  }
-  AdStackOverflowDiagnosis result;
-  result.message = identity_block + disambiguation_block + body +
-                   "Note: kernel state may be inconsistent post-overflow; do not retry the same "
-                   "step without addressing the cause and restarting from a clean state.";
-  // Flag the cache as confirmed-invalid only when the sync rerun positively identified DLPack-bypass
-  // (`required > allocated` for at least one stack with every leaf resolved against the live snapshot).
-  // The diagnose-time evaluator can now resolve ndarray-bound leaves through the captured launch snapshot
-  // (`Device::map`-based reads), so `Unknown` is a rare fallback - either no launch has happened yet
-  // (snapshot not populated), or a `Device::map` call failed, or a leaf type is unsupported by the
-  // diagnose evaluator. None of these correspond to a high-confidence DLPack-bypass, so we conservatively
-  // skip invalidation and let the loud overflow surface to the user instead of silently retrying against
-  // a possibly-broken cache. Quadrants-bug is excluded for the same reason invalidation would be unsafe
-  // (next launch would re-run the same wrong sizer and produce the same wrong bound).
-  result.confirmed_invalid_cache = (cause == Cause::DLPackBypass);
-  return result;
-}
-
-void Program::capture_diagnose_snapshot(const LaunchContextBuilder &ctx) {
-  std::lock_guard<std::mutex> lk(diagnose_snapshot_mutex_);
-  diagnose_snapshot_.data_ptrs.clear();
-  diagnose_snapshot_.dev_alloc_types.clear();
-  diagnose_snapshot_.shapes.clear();
-  // Pull just the data-pointer slot for each arg; the grad-pointer slot is irrelevant to size_expr leaves.
-  for (const auto &kv : ctx.array_ptrs) {
-    if (kv.first.ptr_type == TypeFactory::DATA_PTR_POS_IN_NDARRAY) {
-      diagnose_snapshot_.data_ptrs[kv.first.arg_id] = kv.second;
-    }
-  }
-  diagnose_snapshot_.dev_alloc_types = ctx.device_allocation_type;
-  // Mirror the per-arg shape vectors `LaunchContextBuilder` populated alongside the args-buffer writes. Going
-  // through this side map rather than `args_type->get_element_offset` avoids the spurious "Cannot treat as
-  // TensorType" diagnostics emitted when an axis lookup overruns the actual rank, and keeps the diagnose path
-  // independent of `args_type` lifetime.
-  for (const auto &kv : ctx.ndarray_shapes) {
-    std::vector<int32_t> shape32(kv.second.begin(), kv.second.end());
-    diagnose_snapshot_.shapes[kv.first] = std::move(shape32);
-  }
-  diagnose_snapshot_.valid = true;
-}
-
-const Program::DiagnoseLaunchSnapshot *Program::get_diagnose_snapshot() const {
-  std::lock_guard<std::mutex> lk(diagnose_snapshot_mutex_);
-  return diagnose_snapshot_.valid ? &diagnose_snapshot_ : nullptr;
 }
 
 StreamSemaphore Program::flush() {
